Added binary

+++ src/skyhadoop/Partioner.java
@@ -8,8 +8,8 @@ public class Partioner {
 	
 	
 	public static int gridpartioner(Point p) {
-        double x=p.d[0]/gridX;
-        double y=p.d[1]/gridY;
+        int x=(int)(p.d[0]/gridX);
+        int y=(int)(p.d[1]/gridY);
         
         return (int)(x*sizeX/gridX + y);
